pyproject.toml
[project]
name = "traveler-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Andriy Sydorenko",email = "sidorenkoandrij217@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black (>=25.1.0,<26.0.0)",
    "isort (>=6.0.1,<7.0.0)",
    "ruff (>=0.11.13,<0.12.0)",
    "fastapi[standard] (>=0.115.12,<0.116.0)",
    "alembic (>=1.16.1,<2.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "argon2-cffi (>=25.1.0,<26.0.0)",
    "greenlet (>=3.2.3,<4.0.0)"
]

[tool.isort]
profile = "black"

[tool.ruff]
line-length = 119
exclude = ["**/__init__.py"]
target-version = "py313"

[tool.ruff.format]
quote-style = "double"
line-ending = "auto"
docstring-code-format = true
docstring-code-line-length = 119

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
